recursive publish capability
improves the package publishing process by looking at a "publish" as
just being an effective copy of a package from one registry to another
registry (aka the local registry to the ). A special utility function is
now included in the builder to make this a much more seamless process.

also some fixes for the naming of the "package Info" (meta) in the
deploy manifest, and "meta url" which are two separate things but have
unfortunately very similar names due to how planning has gone.

in the future we will want to get rid of the metadata cache completely
on the cli, but for the time being its fine and what we have now has
really not caused too much problem

still need to add tests and do manual testing