fix(builder): validation of template accesses which depend on `chainId`
or other globals

this basically turned into a refactor because we have to kind of rethink
how we are supplying values to the ChainDefinition. But I think its a
move in the right direction.

Unfortunately refactors like this usually lead to breakage, so probably
best to save this for next release.